navigation in science screen no longer snaps back to first entry
It happened before that in a long database screen the scoll bar
snapped to the top when an entry at the bottom was selected.
The problem was that all entries are added in a loop, each
triggering a redraw of the component. With the logic before
this caused the list to have length 0 at one point which made it
loose its scroll position, because the "value" in "GuiScrollbar"
was set to 0 too. This is now fixed by using a "desired_value"
instead that is maintained even when the number of elements
decreases.